Fix customer delete session cleanup for staff acting.
When deleting an empty customer, clear customer_id on staff acting
wholesale sessions instead of deleting them so internal staff stay
signed in. Buyer wholesale sessions are still removed.

Add PGlite coverage, assert CUST- counter does not rewind after delete,
and reject fractional credit-limit dollars in create/edit forms.

+++ apps/internal/src/components/customer-create-form.tsx
@@ -25,7 +25,7 @@ import { useCanManageStaff } from "../lib/staff-manage";
 const baseCreateCustomerSchema = z.object({
   name: z.string().min(1, "Name is required"),
   terms: z.enum(CUSTOMER_TERMS),
-  creditLimitDollars: z.coerce.number().min(0, "Credit limit is required"),
+  creditLimitDollars: z.coerce.number().int().min(0, "Credit limit is required"),
   wholesaleEmail: z.string().optional(),
   wholesaleDisplayName: z.string().optional(),
 });
